+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
-所有对 Peek 项目的显著更改都将记录在此文件中。
+所有对 PeekView 项目的显著更改都将记录在此文件中。
 
 格式基于 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.0.0/)，
 并且本项目遵循 [语义化版本](https://semver.org/lang/zh-CN/)。
+
+## [0.1.1] - 2026-04-24
+
+### 变更
+
+- 项目重命名为 PeekView
+- CLI 命令从 `peek` 改为 `peekview`
+- 环境变量前缀从 `PEEK_` 改为 `PEEKVIEW_`
+- 数据目录从 `~/.peek` 改为 `~/.peekview`
+- Python 包名从 `peek` 改为 `peekview`
 
 ## [0.1.0] - 2026-04-23